+++ DesignToolbox/DesignToolbox/Pages/Components/Controls/Checkbox/CheckboxItem/Indeterminate/CheckboxItemIndeterminatePage.swift
@@ -49,6 +49,7 @@ private struct CheckboxItemIndeterminateDemo: View {
                                       flipIcon: configurationModel.flipIcon,
                                       isReversed: configurationModel.isReversed,
                                       isError: configurationModel.isError,
+                                      errorText: configurationModel.errorText,
                                       isReadOnly: configurationModel.isReadOnly,
                                       hasDivider: configurationModel.divider)
             .disabled(!configurationModel.enabled)