+++ DesignToolbox/DesignToolbox/Pages/Components/Controls/Checkbox/CheckboxItem/TwoStates/CheckboxItemPage.swift
@@ -49,6 +49,7 @@ private struct CheckboxItemDemo: View {
                          flipIcon: configurationModel.flipIcon,
                          isReversed: configurationModel.isReversed,
                          isError: configurationModel.isError,
+                         errorText: configurationModel.errorText,
                          isReadOnly: configurationModel.isReadOnly,
                          hasDivider: configurationModel.divider)
             .disabled(!configurationModel.enabled)