+++ DesignToolbox/DesignToolbox/Pages/Components/Controls/Switch/SwitchItem/SwitchItemPage.swift
@@ -48,6 +48,7 @@ private struct SwitchItemDemo: View {
                        flipIcon: configurationModel.flipIcon,
                        isReversed: !configurationModel.isReversed,
                        isError: configurationModel.isError,
+                       errorText: configurationModel.errorText,
                        isReadOnly: configurationModel.isReadOnly,
                        hasDivider: configurationModel.divider)
             .disabled(!configurationModel.enabled)